requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.11
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
email-validator==2.0.0 
debugpy==1.8.16

